+++ apps/minifront/src/components/extension-not-installed.tsx
@@ -10,14 +10,14 @@ export const ExtensionNotInstalled = () => {
       <HeadTag />
       <SplashPage title='Welcome to Penumbra'>
         <div className='flex items-center justify-between gap-[1em] text-lg'>
-          To get started, install the Penumbra Chrome extension.
+          To get started, install a Penumbra extension.
           <Button asChild variant='gradient'>
             <a
               href={`https://chrome.google.com/webstore/detail/penumbra-wallet/${CHROME_EXTENSION_ID}`}
               target='_blank'
               rel='noreferrer'
             >
-              Install
+              Install Prax
             </a>
           </Button>
         </div>
